+++ src/Exception/ConvertException.php
@@ -16,9 +16,23 @@ use const PHP_EOL;
  */
 final class ConvertException
 {
-    private const MSG_INTEGRITY_EXCEPTION_1 = 'SQLSTATE[23';
-    private const MGS_INTEGRITY_EXCEPTION_2 = 'ORA-00001: unique constraint';
-    private const MSG_INTEGRITY_EXCEPTION_3 = 'SQLSTATE[HY';
+    private const MSG_CONNECTION_EXCEPTION = 'SQLSTATE[08';
+    private const MSG_INTEGRITY_EXCEPTION = 'SQLSTATE[23';
+    private const MYSQL_RECONNECT_EXCEPTIONS = [
+        'SQLSTATE[HY000]: General error: 2006 ',
+        'SQLSTATE[HY000]: General error: 4031 ',
+    ];
+    private const ORACLE_COMPATIBILITY_EXCEPTIONS = [
+        'ORA-00942:',
+    ];
+    private const ORACLE_INTEGRITY_EXCEPTIONS = [
+        'ORA-00001:',
+        'ORA-01400:',
+        'ORA-01407:',
+        'ORA-02290:',
+        'ORA-02291:',
+        'ORA-02292:',
+    ];
 
     public function __construct(
         private readonly \Exception $e,
@@ -36,13 +50,52 @@ final class ConvertException
 
         $errorInfo = $this->e instanceof PDOException ? $this->e->errorInfo : null;
 
-        return match (
-            str_contains($message, self::MSG_INTEGRITY_EXCEPTION_1)
-            || str_contains($message, self::MGS_INTEGRITY_EXCEPTION_2)
-            || str_contains($message, self::MSG_INTEGRITY_EXCEPTION_3)
+        if (
+            str_contains($message, self::MSG_INTEGRITY_EXCEPTION)
+            || $this->isMysqlReconnectException($message)
+            || $this->isOracleCompatibilityException($message)
+            || $this->isOracleIntegrityException($message)
         ) {
-            true => new IntegrityException($message, $errorInfo, $this->e),
-            default => new Exception($message, $errorInfo, $this->e),
-        };
+            return new IntegrityException($message, $errorInfo, $this->e);
+        }
+
+        if (str_contains($message, self::MSG_CONNECTION_EXCEPTION)) {
+            return new ConnectionException($message, $errorInfo, $this->e);
+        }
+
+        return new Exception($message, $errorInfo, $this->e);
+    }
+
+    private function isMysqlReconnectException(string $message): bool
+    {
+        foreach (self::MYSQL_RECONNECT_EXCEPTIONS as $mysqlReconnectException) {
+            if (str_contains($message, $mysqlReconnectException)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    private function isOracleCompatibilityException(string $message): bool
+    {
+        foreach (self::ORACLE_COMPATIBILITY_EXCEPTIONS as $oracleCompatibilityException) {
+            if (str_contains($message, $oracleCompatibilityException)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    private function isOracleIntegrityException(string $message): bool
+    {
+        foreach (self::ORACLE_INTEGRITY_EXCEPTIONS as $oracleIntegrityException) {
+            if (str_contains($message, $oracleIntegrityException)) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }